# A03_OpenAPI_2206081925.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Marketplace API",
    "description": "Dokumentasi API untuk arsitektur microservices"
  },
  "host": "localhost:8000",
  "basePath": "/api",
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "paths": {
...
    "/order": {
      "post": {
        "summary": "Pengguna melakukan pemesanan (checkout cart)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["shipping_provider"],
              "properties": {
                "shipping_provider": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Order berhasil dibuat"
          }
        }
      }
    },
    "/cart": {
      "get": {
        "summary": "Ambil semua item dalam cart pengguna",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List item cart"
          }
        }
      }
    },
...
  }
}